# api/openapi.3.1.projection.yaml
...
openapi: 3.1.0
info:
  title: Bedrock API
  version: 2.0.0
  description: |
    Bedrock 2.0 HTTP API. Envelope `{ code, message, data? }`.
    Pagination data shape: `{ items, total, page, page_size, total_pages }`.
servers:
  - url: /api/v1
    description: Relative to the Bedrock host
security:
  - bearerAuth: []
...
paths:
...
  /repositories:
    get:
      tags: [Repositories]
      summary: List repositories
      operationId: repositoriesList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema: {type: string}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/RepositoryPage"
        "403":
          $ref: "#/components/responses/Error"
    post:
      tags: [Repositories]
      summary: Create repository
      operationId: repositoriesCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Repository"
        "403":
          $ref: "#/components/responses/Error"
  /repositories/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Repositories]
      summary: Get repository
      operationId: repositoriesGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Repository"
        "404":
          $ref: "#/components/responses/Error"
    put:
      tags: [Repositories]
      summary: Update repository
      operationId: repositoriesUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Repository"
        "403":
          $ref: "#/components/responses/Error"
    delete:
      tags: [Repositories]
      summary: Delete repository
      operationId: repositoriesDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          $ref: "#/components/responses/Error"
  /repositories/{id}/branches:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Repositories]
      summary: List remote branches
      operationId: repositoriesBranches
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        type: object
                        properties:
                          items:
                            type: array
                            items: {type: string}
  /repositories/{id}/test:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Repositories]
      summary: Test fetch / list branches
      operationId: repositoriesTest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /repositories/{id}/webhook-secret:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Repositories]
      summary: View webhook secret and URL
      operationId: repositoriesWebhookSecretGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
  /repositories/{id}/webhook-secret/rotate:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Repositories]
      summary: Rotate webhook secret
      operationId: repositoriesWebhookSecretRotate
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /servers:
    get:
      tags: [Servers]
      summary: List servers
      operationId: serversList
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PageSize"
        - name: keyword
          in: query
          schema: {type: string}
        - name: tag
          in: query
          schema: {type: string}
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/ServerPage"
    post:
      tags: [Servers]
      summary: Create server
      operationId: serversCreate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerCreateRequest"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Server"
  /servers/{id}:
    parameters:
      - $ref: "#/components/parameters/Id"
    get:
      tags: [Servers]
      summary: Get server
      operationId: serversGet
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Server"
    put:
      tags: [Servers]
      summary: Update server
      operationId: serversUpdate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ServerUpdateRequest"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/SuccessEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/Server"
    delete:
      tags: [Servers]
      summary: Delete server
      operationId: serversDelete
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "409":
          $ref: "#/components/responses/Error"
  /servers/{id}/test:
    parameters:
      - $ref: "#/components/parameters/Id"
    post:
      tags: [Servers]
      summary: Test server connection
      operationId: serversTest
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
...
  /webhook/repos/{repository_id}/{secret}:
    parameters:
      - name: repository_id
        in: path
        required: true
        schema: {type: integer}
      - name: secret
        in: path
        required: true
        schema: {type: string}
    post:
      tags: [Webhook]
      summary: Receive repository webhook (signature preferred; URL secret fallback; delivery dedup)
      security: []
      operationId: webhookReceive
      parameters:
        - name: build_job_id
          in: query
          schema: {type: integer}
          description: Optional narrow to a single BuildJob
      responses:
        "202":
          description: Accepted (may be duplicate with triggered=0)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessEnvelope"
        "401":
          $ref: "#/components/responses/Error"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema:
        type: integer
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSize:
      name: page_size
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
  responses:
    Error:
      description: Error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    SuccessEnvelope:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
        message:
          type: string
        request_id:
          type: string
...
    Page:
      type: object
      required: [items, total, page, page_size, total_pages]
      properties:
        items:
          type: array
          items: {}
        total:
          type: integer
          format: int64
        page:
          type: integer
        page_size:
          type: integer
        total_pages:
          type: integer
...
    Repository:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        description: {type: string}
        tags: {type: string}
        repo_url: {type: string}
        default_branch: {type: string}
        auth_type: {type: string, enum: [none, credential]}
        credential_id: {type: integer, nullable: true}
        webhook_secret: {type: string, description: "Only present on secret view/rotate"}
        webhook_type: {type: string}
        webhook_ref_path: {type: string}
        webhook_commit_path: {type: string}
        webhook_message_path: {type: string}
        created_by: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    RepositoryCreateRequest:
      type: object
      required: [name, repo_url]
      properties:
        name: {type: string}
        description: {type: string}
        tags: {type: string}
        repo_url: {type: string}
        default_branch: {type: string}
        auth_type: {type: string}
        credential_id: {type: integer, nullable: true}
        webhook_type: {type: string}
        webhook_ref_path: {type: string}
        webhook_commit_path: {type: string}
        webhook_message_path: {type: string}
    RepositoryUpdateRequest:
      type: object
      properties:
        name: {type: string}
        description: {type: string}
        tags: {type: string}
        repo_url: {type: string}
        default_branch: {type: string}
        auth_type: {type: string}
        credential_id: {type: integer, nullable: true}
        clear_credential: {type: boolean}
        webhook_type: {type: string}
        webhook_ref_path: {type: string}
        webhook_commit_path: {type: string}
        webhook_message_path: {type: string}
    RepositoryPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Repository"
...
    Server:
      type: object
      properties:
        id: {type: integer}
        name: {type: string}
        host: {type: string}
        port: {type: integer}
        os_type: {type: string}
        username: {type: string}
        auth_type: {type: string}
        credential_id: {type: integer, nullable: true}
        agent_url: {type: string}
        agent_credential_id: {type: integer, nullable: true}
        description: {type: string}
        tags: {type: string}
        status: {type: string}
        created_by: {type: integer}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    ServerCreateRequest:
      type: object
      required: [name]
      properties:
        name: {type: string}
        host: {type: string}
        port: {type: integer}
        os_type: {type: string}
        username: {type: string}
        auth_type: {type: string}
        credential_id: {type: integer, nullable: true}
        agent_url: {type: string}
        agent_credential_id: {type: integer, nullable: true}
        description: {type: string}
        tags: {type: string}
    ServerUpdateRequest:
      type: object
      properties:
        name: {type: string}
        host: {type: string}
        port: {type: integer}
        os_type: {type: string}
        username: {type: string}
        auth_type: {type: string}
        credential_id: {type: integer, nullable: true}
        clear_credential: {type: boolean}
        agent_url: {type: string}
        agent_credential_id: {type: integer, nullable: true}
        clear_agent_credential: {type: boolean}
        description: {type: string}
        tags: {type: string}
    ServerPage:
      allOf:
        - $ref: "#/components/schemas/Page"
        - type: object
          properties:
            items:
              type: array
              items:
                $ref: "#/components/schemas/Server"
...